eleminate cime_util in progress
